init Commit

+++ waesworks-differ-for-testers/src/test/java/org/waes/differ/utils/Helpers/Common.java
@@ -4,7 +4,7 @@ import org.apache.commons.lang3.RandomUtils;
 
 public class Common {
 
-    public static int getRandomDigit() {
+    public static long getRandomDigit() {
         return RandomUtils.nextInt(100, 100000);
     }
